Fix: move isSend to src/internals/tact/util.ts Fix: callgraph id Chore: add comments to Flag const for CGnode
